// web/src/api-access/generated/models/schedulingCalendarRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Unified.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface SchedulingCalendarRequest {
  startDate: string;
  endDate: string;
  /** @nullable */
  timeZoneId?: string | null;
  /** @nullable */
  locationId?: number | null;
  /** @nullable */
  userIds?: string[] | null;
}